DB保存するための準備
