+++ app/assets/javascripts/application.js
@@ -2,42 +2,6 @@
 //= require popper
 //= require bootstrap
 
-// mizu--------------------------------------------------
-window.addEventListener('load', function(){
-
-  const pullDownButton = document.getElementById("lists")
-
-  pullDownButton.addEventListener('mouseover', function(){
-    this.setAttribute("style", "background: url('/images/one.png');")
-  })
-
-  pullDownButton.addEventListener('mouseout', function(){
-    this.setAttribute("style", "background: url('/images/one.png');")
-  })
-
-  let count = 0;
-  pullDownButton.addEventListener('click', function() {
-    count += 1
-    if (count == 1) {
-      pullDownButton.setAttribute("style", "background: url('/images/two.png');")
-    } else if (count == 2) {
-      pullDownButton.setAttribute("style", "background: url('/images/three.png');") 
-    } else if (count == 3) {
-      pullDownButton.setAttribute("style", "background: url('/images/four.png');")
-    } else if (count == 4) {
-      pullDownButton.setAttribute("style", "background: url('/images/five.png');")
-    } else if (count == 5) {
-      pullDownButton.setAttribute("style", "background: url('/images/six.png');")
-    } else if (count == 6) {
-      count = 0
-    } else {
-      pullDownButton.setAttribute("style", "background: url('/images/six.png');")
-    }
-
-  })
-})
-// mizu-----------------------------------//
-
 // timer----------------------------------
 window.onload = initialize;
 function initialize() {
@@ -103,5 +67,21 @@ function stopwatchClear() {
 }
 // -->
 // timer------------------------------//
+// timer2------------------------------
+window.addEventListener('load', function(){
 
+  const pullDownButton = document.getElementById("stopwatchRecord")
+  const hour = document.getElementById("hour")
+  const minute = document.getElementById("minute")
+  const second = document.getElementById("second")
+
+  pullDownButton.addEventListener('click', function() {
+    hour.innerHTML = $stopwatchHour;
+    minute.innerHTML = $stopwatchMinute;
+    second.innerHTML = $stopwatchSecond;
+    
+
+  })
+})
+//timer2-----------------------------//
