+++ app/controllers/meals_controller.rb
@@ -1,5 +1,10 @@
 class MealsController < ApplicationController
   
   def new
+    @meal = Meal.new
+  end
+
+  def create
+    redirect_to record_index_path
   end
 end
