+++ app/views/record/top.html.erb
@@ -2,7 +2,7 @@
   <div class="container-fluid jumbotron-container">
     <h1 class="site-name" class="shadow-sm" ><span class="top-html">生活習慣</span>を<span class="top-html">記録</span>しよう</h1>
     <% if user_signed_in? %>
-      <p><%= link_to 'Go record', record_index_path, class: "btn btn-black"%></p>
+      <p><%= link_to 'Go record', record_index_path, data: {"turbolinks" => false}, class: "btn btn-black"%></p>
     <% else %>
       <%= link_to 'Sign in', new_user_session_path, class: "btn btn-black"%>
       <%= link_to 'Guest', record_guest_sign_in_path, method: :post, class: "btn btn-black guest-btn"%>
