+++ app/views/sleeps/new.html.erb
@@ -1,30 +1,29 @@
-<h2>---</h2>
-<p>---</p>
-
-<div class="upper-content">
-  <h1 class="heading-name">睡眠</h1>
-  <div class="meal-record">
-    <div id="stopwatchWrapper">
-  <div id="stopwatch">
-    <span id="stopwatchHour" class="count">00</span>
-      <span>:</span>
-      <span id="stopwatchMinute" class="count">00</span>
-      <span>:</span>
-      <span id="stopwatchSecond" class="count">00</span>
-      <span>:</span>
-      <span id="stopwatchMillisecond" class="count">000</span>
+<div class="jumbotron jumbotron-newpage mt-4">
+  <div class="upper-content">
+    <h1 class="heading-name mt-5">睡眠</h1>
+    <div class="meal-record">
+      <div id="stopwatchWrapper">
+    <div id="stopwatch" class="watch">
+      <span id="stopwatchHour" class="count">00</span>
+        <span class="count">:</span>
+        <span id="stopwatchMinute" class="count">00</span>
+        <span class="count">:</span>
+        <span id="stopwatchSecond" class="count">00</span>
+        <span class="count">.</span>
+        <span id="stopwatchMillisecond" class="second">000</span>
+    </div>
+    <div id="stopwatchControl">
+      <button id="stopwatchStartAndStop">スタート</button>
+      <button id="stopwatchClear">クリア</button>
+    </div>
   </div>
-  <div id="stopwatchControl">
-    <button id="stopwatchStartAndStop">スタート</button>
-    <button id="stopwatchClear">クリア</button>
-  </div>
-</div>
-  </div>
-  <div class="memo">
-    <div>メモ</div>
-    <textarea class="memo-input"></textarea>
-  </div>
-  <div class="exit">
-    <input type="submit" value="終了" class="exit-sub">
+    </div>
+    <div class="memo">
+      <div>メモ</div>
+      <textarea class="memo-input"></textarea>
+    </div>
+    <div class="exit">
+      <input type="submit" value="終了" class="exit-sub">
+    </div>
   </div>
 </div>
