+++ db/migrate/20200919100651_rename_time_column_to_meals.rb
@@ -1,0 +1,5 @@
+class RenameTimeColumnToMeals < ActiveRecord::Migration[6.0]
+  def change
+    rename_column :meals, :time, :hour
+  end
+end
