+++ db/migrate/20200919102119_add_minute_to_meals.rb
@@ -1,0 +1,5 @@
+class AddMinuteToMeals < ActiveRecord::Migration[6.0]
+  def change
+    add_column :meals, :minute, :integer
+  end
+end
